moved axios calls to App
