+++ src/App.js
@@ -1,26 +1,80 @@
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { BrowserRouter as Router, Link, Route} from 'react-router-dom';
 import './App.css';
 import Videos from './components/Videos';
 import Customers from './components/Customers'
+import axios from 'axios'
+
+const BASE_URL = 'http://localhost:3000';
 
 const App = () => {
 
   const [selectedVideo, setSelectedVideo] = useState('');
+  const [videos, setVideos] = useState([]);
+  const [customers, setCustomers] = useState([]);
+  const [errorMessage, setErrorMessage] = useState([]);
 
   const setVideo = (videoTitle) => {
     setSelectedVideo(videoTitle)
   }
+
+  useEffect(() => {
+    axios.get(BASE_URL + '/videos')
+    .then((response) => {
+      const apiVideos = response.data.map((apiVideo) => {
+        return ({
+          id: apiVideo.id,
+          title: apiVideo.title,
+          overview: apiVideo.overview,
+          releaseDate: apiVideo.release_date,
+          imageUrl: apiVideo.image_url,
+          externalID: apiVideo.externalID
+        })
+      })
+      setVideos(apiVideos)
+    })
+    .catch((error) => {
+      setErrorMessage(error.message)
+    })
+  }, [])
+
+
+  useEffect(() => {
+    axios.get(BASE_URL + '/customers')
+    .then((response) => {
+    const apiCustomers = response.data.map((apiCustomer) => {
+        return ({
+        id: apiCustomer.id,
+        name: apiCustomer.name,
+        registeredAt: apiCustomer.registered_at,
+        address: apiCustomer.address,
+        city: apiCustomer.city, 
+        state: apiCustomer.state,
+        postalCode: apiCustomer.postal_code,
+        phone: apiCustomer.phone,
+        accountCredit: apiCustomer.account_credit,
+        videosCheckedOutCount: apiCustomer.videos_checked_out_count
+        })
+    })
+    setCustomers(apiCustomers)
+    })
+    .catch((error) => {
+        setErrorMessage(error.message)
+    })
+}, [])
   
   return (
-    <Router>
+    <div>
+      {errorMessage ? <div><h2>{errorMessage}</h2></div> : ''}
+      <Router>
       <Link to='/'>Home</Link>
       <Link to='/customers'>Customers</Link>
       <Link to='/library'>Videos</Link>
       <Route path='/' render={() => selectedVideo} />
-      <Route path='/customers' component={() => <Customers url='http://localhost:3000' />}/>
-      <Route path='/library' component={() => <Videos url='http://localhost:3000' onClickCallback={setVideo}/>}/>
+      <Route path='/customers' component={() => <Customers customers={customers} />}/>
+      <Route path='/library' component={() => <Videos videos={videos} onClickCallback={setVideo}/>}/>
     </Router>
+    </div>
   );
 }
